+++ src/broadcast.rs
@@ -1,31 +1,5 @@
-use crate::netaddr::NetAddr;
-use crate::netv4addr::Netv4Addr;
-use std::net::{IpAddr, Ipv4Addr};
-
 pub trait Broadcast {
 	type Output;
 
 	fn broadcast(&self) -> Self::Output;
-}
-
-impl Broadcast for Netv4Addr {
-	type Output = Ipv4Addr;
-
-	fn broadcast(&self) -> Ipv4Addr {
-		let netmask: u32 = self.mask().clone().into();
-		let network: u32 = self.addr().clone().into();
-		let broadcast: u32 = network | !netmask;
-		broadcast.into()
-	}
-}
-
-impl Broadcast for NetAddr {
-	type Output = Option<IpAddr>;
-
-	fn broadcast(&self) -> Self::Output {
-		match self {
-			NetAddr::V4(netaddr) => Some(IpAddr::from(netaddr.broadcast())),
-			_ => None,
-		}
-	}
 }
